Keep Hermes support within Swift budgets
