+++ Sources/SessionTranscriptTypes.swift
@@ -1,0 +1,59 @@
+import SwiftUI
+
+struct SessionTranscriptTurn: Identifiable, Equatable, Sendable {
+    let id: Int
+    let role: SessionTranscriptRole
+    let text: String
+}
+
+enum SessionTranscriptRole: Equatable, Sendable {
+    case user
+    case assistant
+    case system
+    case tool
+    case event
+
+    var label: String {
+        switch self {
+        case .user:
+            return String(localized: "sessionIndex.preview.role.user", defaultValue: "You")
+        case .assistant:
+            return String(localized: "sessionIndex.preview.role.assistant", defaultValue: "Agent")
+        case .system:
+            return String(localized: "sessionIndex.preview.role.system", defaultValue: "System")
+        case .tool:
+            return String(localized: "sessionIndex.preview.role.tool", defaultValue: "Tool")
+        case .event:
+            return String(localized: "sessionIndex.preview.role.event", defaultValue: "Event")
+        }
+    }
+
+    var foregroundColor: Color {
+        switch self {
+        case .user: return .accentColor
+        case .assistant: return .green
+        case .system: return .secondary
+        case .tool: return .orange
+        case .event: return .secondary
+        }
+    }
+
+    var backgroundColor: Color {
+        switch self {
+        case .user: return Color.accentColor.opacity(0.035)
+        case .assistant: return Color.green.opacity(0.035)
+        case .system: return Color.primary.opacity(0.025)
+        case .tool: return Color.orange.opacity(0.035)
+        case .event: return Color.primary.opacity(0.02)
+        }
+    }
+
+    var bodyFont: Font {
+        switch self {
+        case .tool, .system:
+            return .system(size: 11, design: .monospaced)
+        case .user, .assistant, .event:
+            return .system(size: 12)
+        }
+    }
+}
